+++ src/commands/general/serverid.js
@@ -1,0 +1,31 @@
+/**
+ * Discord Welcome-Bot
+ * Copyright (c) 2021 The Welcome-Bot Team and Contributors
+ * Licensed under Lesser General Public License v2.1 (LGPl-2.1 - https://opensource.org/licenses/lgpl-2.1.php)
+ */
+//eslint-disable-next-line no-unused-vars
+const { Embed, Command } = require("../../classes");
+module.exports = class CMD extends Command {
+    constructor(client) {
+        super(
+            {
+                name: "serverid",
+                aliases: ["sid", "guild-id"],
+                memberPerms: [],
+                botPerms: [],
+                requirements: {
+                    guildOnly: true,
+                },
+                disabled: false,
+                cooldown: 5,
+                category: "General",
+            },
+            client
+        );
+    }
+
+    //eslint-disable-next-line no-unused-vars
+    execute({ message, args, guildDB, userDB }, t) {
+        message.channel.send(`${message.guild.id}`);
+    }
+};