+++ src/commands/general/userid.js
@@ -1,0 +1,35 @@
+/**
+ * Discord Welcome-Bot
+ * Copyright (c) 2021 The Welcome-Bot Team and Contributors
+ * Licensed under Lesser General Public License v2.1 (LGPl-2.1 - https://opensource.org/licenses/lgpl-2.1.php)
+ */
+//eslint-disable-next-line no-unused-vars
+const { Embed, Command } = require("../../classes");
+module.exports = class CMD extends Command {
+    constructor(client) {
+        super(
+            {
+                name: "userid",
+                aliases: ["uid", "user-id"],
+                memberPerms: [],
+                botPerms: [],
+                requirements: {
+                    args: true,
+                },
+                disabled: false,
+                cooldown: 5,
+                category: "General",
+            },
+            client
+        );
+    }
+
+    //eslint-disable-next-line no-unused-vars
+    async execute({ message, args, guildDB, userDB }, t) {
+        const user = await this.getUserFromIdOrMention(args[0]);
+        if (!user) {
+            return message.reply(t("errors:invalidUser"));
+        }
+        message.channel.send(`${user.id}`);
+    }
+};